+++ python/sglang/test/ci/ci_utils.py
@@ -7,8 +7,8 @@ import time
 from dataclasses import dataclass
 from typing import Callable, List, Optional, Union
 
+from sglang.srt.debug_utils import cuda_coredump
 from sglang.srt.utils.common import kill_process_tree
-from sglang.test.ci import cuda_coredump
 from sglang.test.ci.ci_register import CIRegistry
 
 # Configure logger to output to stdout
